--- src/pages/buttons.js
@@ -1,7 +1,0 @@
-import Spline from '@splinetool/react-spline';
-
-export default function Buttons() {
-  return (
-    <Spline scene="https://prod.spline.design/zvu6gCvwd35arEyW/scene.splinecode" />
-  );
-}
