+++ src/pages/home.js
@@ -1,7 +1,7 @@
 import { useNavigate } from "react-router-dom";
 import { useState, useEffect } from "react";
-import Buttons from "./buttons";
 import '../App.css';
+import Background from "../components/Background"
 
 
 const Home = () => {
@@ -9,14 +9,11 @@ const Home = () => {
 
   const [aboutOpen, setAboutOpen] = useState(false);
 
-  let backgroundImage = "https://cdn.glitch.global/4a3253fe-3104-44d3-8a31-8f43e023369d/circles.gif?v=1702433383009"
-  let yonderFrameImage = "https://cdn.glitch.global/4a3253fe-3104-44d3-8a31-8f43e023369d/yonderGardenFrameFg.png?v=1702515498646"
-  let yonderGrassImage = "https://cdn.glitch.global/4a3253fe-3104-44d3-8a31-8f43e023369d/grassLandscape.png?v=1702081026940"
-  let titleImage = "https://cdn.glitch.global/4a3253fe-3104-44d3-8a31-8f43e023369d/yonderGardenFrameBackgroundWords.png?v=1702515532263"
   let episodesMenuImage = "https://i.imgur.com/tHMXkUr.png"
   let premiumMenuImage = "https://i.imgur.com/UHet8Vq.png"
   let merchMenuImage = "https://i.imgur.com/EEKI1Td.png"
   let aboutMenuImage = "https://i.imgur.com/ZjdrIPQ.png"
+
 
   let navigate = useNavigate();
 
@@ -27,6 +24,19 @@ const Home = () => {
   useEffect(() => {
     console.log(aboutOpen)
   },[aboutOpen])
+
+  const SideMenu = () => {
+    //TODO: Add title, fix button
+    return (
+      <div className="side-menu">
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+        <div className="side-menu-button"></div>
+      </div>
+    );
+  }
 
   const DisplayInfo = () => {
     return (
@@ -42,50 +52,34 @@ const Home = () => {
     );
   };
 
-  const DisplayBackground = () => {
+  // const DisplayBackground = () => {
 
-    return (
-      <div>
-          <img className="center-home layer-51" src={titleImage} alt="" />
+  //   return (
+  //     <div>
+  //       <img className="center-home layer-51" src={titleImage} alt="" />
+  //       <div className='bind-bg'>
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //         <img className="bg-align" src={backgroundImage} alt="" />
+  //       </div>
+  //       <div className='bind-bg'>
+  //         <img className="bg-align" src={yonderGrassImage} alt="" />
+  //         <img className="bg-align" src={yonderGrassImage} alt="" />
+  //         <img className="bg-align" src={yonderGrassImage} alt="" /> 
+  //         <img className="bg-align" src={yonderGrassImage} alt="" />
+  //         <img className="bg-align" src={yonderGrassImage} alt="" /> 
+  //         <img className="bg-align" src={yonderGrassImage} alt="" />
+  //         <img className="bg-align" src={yonderGrassImage} alt="" /> 
+  //       </div>
 
-          {/* <img className="center-home layer-50 " src={yonderFrameImage} alt="" /> */}
-
-        <div className='bind-bg'>
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-          <img className="test-bg-align" src={backgroundImage} alt="" />
-        </div>
-        <div className='bind-bg'>
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-          <img className="test-bg-align" src={yonderGrassImage} alt="" />
-        </div>
-
-          <img className="center-home layer-50 " src={yonderFrameImage} alt="" />
-
-          {/* <img className="center-home layer-1 offset-bg-right" src={yonderGrassImage} alt="" /> */}
-          {/* <img className="center-home layer-1 offset-bg-right" src={yonderGrassImage} alt="" /> */}
-
-
-          <img className="center-home layer-1 " src={yonderGrassImage} alt="" />
-          {/* <img className="center-home layer-1" src={backgroundImage} alt="" /> */}
-
-          {/* <img className="center-home layer-0 offset-bg-right" src={backgroundImage} alt="" /> */}
-          {/* <img className="center-home layer-1 offset-bg-right" src={yonderGrassImage} alt="" /> */}
-
-          {/* <img className="center-home layer-0 offset-bg-left" src={backgroundImage} alt="" />
-          <img className="center-home layer-1 offset-bg-left" src={yonderGrassImage} alt="" /> */}
-      </div>
-      );
-  }
+  //       <img className="center-home layer-50 " src={yonderFrameImage} alt="" />
+  //     </div>
+  //     );
+  // }
 
   const DisplayMenu = () => {
 
@@ -102,8 +96,11 @@ const Home = () => {
   return (
       <header>
         <body >
-        <DisplayBackground />
-        {aboutOpen ? <DisplayInfo/> : <DisplayMenu />}
+        <SideMenu />
+          <div>
+            <Background />
+            {aboutOpen ? <DisplayInfo/> : <DisplayMenu />}
+          </div>
         </body>
       </header>
